aic-24-BE/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asonic==2.0.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
fastapi==0.112.2
filelock==3.15.4
fsspec==2024.6.1
ftfy==6.2.3
fuzzywuzzy==0.18.0
h11==0.14.0
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
open_clip_torch==2.26.1
packaging==24.1
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
sniffio==1.3.1
starlette==0.38.2
sympy==1.13.2
timm==1.0.9
tokenizers==0.19.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.2
uvicorn==0.30.6
wcwidth==0.2.13
